Improve the main menu UI
This commit improves the simple UI we had for the main menu by
using images with icons, organizing them into a grid layout,
and using cards for each button.

It's better to keep improving the UI alongside implementing the
rest of the project and using the improved UI version *if* it's
a viable option (as it's not a priority).