Adds the AlertDialogs that display after hitting the notification
Now that we have a notification and the dialogs, it'll prompt the user
for their mood after they hit the notification, move to the SeekBar
containing dialog, then finally say that it saved the data.

All it has to do now is collect that information and save it to a
database. The UI requires some improvement (e.g. animations, clean up)
and the code requires cleaning.